+++ back_local/src/main/java/com/portfolioweb/aj/Repositorio/RTipoUbicacion.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RTipoUbicacion extends JpaRepository<TipoUbicacion, Long> {
+public interface RTipoUbicacion extends JpaRepository<TipoUbicacion, Integer> {
 
     Optional<TipoUbicacion> findByNombre(String nombre);
